Fix: create hash in capture requests with unknown databases
If you are using an unknown database, no hash was created during the capture request.
This caused the following error to occur:

Error in paste0(type, "-", hash, ".R") :
 cannot coerce type 'closure' to vector of type 'character'

Now in the default case a hash is created over with toString on the object.
There is a risk that files are created twice for the same queries that are called multiple times.
Furthermore, existing files could be overwritten if different queries are executed
but the object returns the same toString.
This solution should serve as a workaround. A more stable solution would be even better.